// src/generated/models/EnPrueba_ProyectosVinculacionModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

export interface AuthorValue {
  "@odata.type": string;
  Claims: string;
  DisplayName: string;
  Email: string;
  Picture: string;
  Department: string;
  JobTitle: string;
}

export interface EditorValue {
  "@odata.type": string;
  Claims: string;
  DisplayName: string;
  Email: string;
  Picture: string;
  Department: string;
  JobTitle: string;
}

export interface EnPrueba_ProyectosVinculacionBase {
  ID?: number;
  Title?: string;
  codigoProyecto?: string;
  nombreProyecto?: string;
  estado?: string;
  coordinadorResponsable?: string;
  correoCoordinador?: string;
  unidadResponsable?: string;
  fechaGuardado?: string;
  datosCompletos?: string;
  OData__ColorTag?: string;
  ComplianceAssetId?: string;
  Modified?: string;
  Created?: string;
  "Author#Claims"?: string;
  "Editor#Claims"?: string;
  "{Identifier}"?: string;
  "{IsFolder}"?: boolean;
  "{Thumbnail}"?: {
    "@odata.type"?: string;
    // Dirección URL de la miniatura pequeña y muy comprimida del elemento, si está disponible.
    Small?: string;
    // Dirección URL de la miniatura de tamaño estándar del elemento, si está disponible.
    Medium?: string;
    // Dirección URL de la miniatura de mayor tamaño del elemento, si está disponible.
    Large?: string;
  };
  "{Link}"?: string;
  "{Name}"?: string;
  "{FilenameWithExtension}"?: string;
  "{Path}"?: string;
  "{FullPath}"?: string;
  "{ModerationStatus}"?: string;
  "{ModerationComment}"?: string;
  "{ContentType}#Id"?: string;
  "{ContentType}"?: {
    "@odata.type"?: string;
    Id?: string;
    Name?: string;
  };
  "{HasAttachments}"?: boolean;
  "{Attachments}@odata.type"?: string;
  "{Attachments}"?: {
   "@odata.type"?: string;
   Id?: string;
   AbsoluteUri?: string;
   DisplayName?: string;
 }[];
  "{VersionNumber}"?: string;
  "{TriggerWindowStartToken}"?: string;
  "{TriggerWindowEndToken}"?: string;
}

export interface EnPrueba_ProyectosVinculacionRead extends EnPrueba_ProyectosVinculacionBase {
  Author?: AuthorValue;
  Editor?: EditorValue;
}

export interface EnPrueba_ProyectosVinculacionWrite extends EnPrueba_ProyectosVinculacionBase {
  Author?: AuthorValue;
  Editor?: EditorValue;
}